requirements.txt
streamlit
plotly
trimesh
numpy
pandas
openai
python-dotenv
scipy
